changed types of pose_time_offset and tracking_frame_offset to float added better parsing of numeric types in the ui

+++ alvr/common/src/data/settings.rs
@@ -180,7 +180,7 @@ pub struct ControllersDesc {
     #[schema(advanced)]
     pub recenter_button: u32,
 
-    pub pose_time_offset: u32,
+    pub pose_time_offset: f32,
 
     #[schema(advanced)]
     pub position_offset_left: [f32; 3],
@@ -218,7 +218,7 @@ pub struct HeadsetDesc {
     #[schema(advanced)]
     pub registered_device_type: String,
 
-    pub tracking_frame_offset: i32,
+    pub tracking_frame_offset: f32,
 
     #[schema(advanced)]
     pub position_offset: [f32; 3],
@@ -373,7 +373,7 @@ pub fn settings_cache_default() -> SettingsDefault {
             manufacturer_name: "Oculus".into(),
             render_model_name: "generic_hmd".into(),
             registered_device_type: "oculus/1WMGH000XX0000".into(),
-            tracking_frame_offset: 0,
+            tracking_frame_offset: 0.,
             position_offset: [0., 0., 0.],
             use_tracking_reference: false,
             force_3dof: false,
@@ -395,7 +395,7 @@ pub fn settings_cache_default() -> SettingsDefault {
                     trackpad_touch_mode: 29,
                     back_mode: 0,
                     recenter_button: 0,
-                    pose_time_offset: 0,
+                    pose_time_offset: 0.,
                     position_offset_left: [-0.007, 0.005, -0.053],
                     rotation_offset_left: [36., 0., 0.],
                     haptics_intensity: 1.,
